refactor(renderer): move HUD Draw pass into renderer_draw.cpp
renderer.cpp was 2678 lines, ~2300 of which was the monolithic Renderer::Draw HUD/overlay pass. Move Draw plus its private font/text helpers (EnsureEditorFont/Sm, DrawFontText/Sm, the g_editorFont* statics) and UnprojectCursorToWorld into renderer_draw.cpp; shared includes go to renderer_internal.h. renderer.cpp core drops to 170 lines (ctor/Init/Shutdown, level setup, VB/IB mapping, SetupUBOMats). Pure code motion: builds clean, runtime-verified (level 2, 558 objects, HUD renders each frame, no errors).
